drm: msm: oplus: Skip pcc on fingerprint_pressed instead
Instead of turning off color processing whenever fingerprint_mode is
true, turn it off whenever fingerprint_pressed is true. This allows
Night Light, LiveDisplay adjustments, etc. to still be applied whenever
the UDFPS circle is on the screen, but allows for turning those off when
the user is actually pressing on the UDFPS circle.

Change-Id: I68810fb2690b345f64368b5939eba4d10ea3b771